docs: Generate CLI documentation
