+++ cmd/cli/command/get-study.go
@@ -1,4 +1,18 @@
-package main
+// Copyright 2018 The Kubeflow Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package command
 
 import (
 	"context"
@@ -9,9 +23,10 @@ import (
 	"text/tabwriter"
 	"unicode/utf8"
 
-	"github.com/kubeflow/katib/pkg/api"
 	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
+
+	"github.com/kubeflow/katib/pkg/api"
 )
 
 //NewCommandGetStudy generate get studies cmd
