+++ cmd/cli/command/get.go
@@ -1,0 +1,37 @@
+// Copyright 2018 The Kubeflow Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package command
+
+import (
+	"github.com/spf13/cobra"
+)
+
+//NewCommandGet generate get cmd
+func NewCommandGet() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "get",
+		Short: "Display one or many resources",
+		Long:  `list of resorces comannd can display includes: studies, study, trials, trial, models, model`,
+	}
+
+	//set local flag
+
+	//add subcommand
+	cmd.AddCommand(NewCommandGetStudy())
+	//	cmd.AddCommand(NewCommandGetTrial())
+	cmd.AddCommand(NewCommandGetModel())
+
+	return cmd
+}
