+++ cmd/cli/command/push.go
@@ -1,0 +1,32 @@
+// Copyright 2018 The Kubeflow Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package command
+
+import (
+	"github.com/spf13/cobra"
+)
+
+//NewCommandPush generate run cmd
+func NewCommandPush() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "push",
+		Short: "Push a resource from a file or from stdin.",
+		Long:  `YAML or JSON formats are accepted.`,
+	}
+
+	cmd.AddCommand(NewCommandPushStudy())
+
+	return cmd
+}
