+++ cmd/cli/katibctl.go
@@ -3,12 +3,14 @@ package main
 import (
 	"fmt"
 	"os"
+
+	"github.com/kubeflow/katib/cmd/cli/command"
 )
 
 //Entry point
 func main() {
 	//init command
-	katibctl := NewRootCommand()
+	katibctl := command.NewRootCommand()
 	if err := katibctl.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
