--- cmd/cli/type.go
@@ -1,10 +1,0 @@
-package main
-
-import (
-	"github.com/kubeflow/katib/pkg/api"
-)
-
-type StudyData struct {
-	StudyConf *api.StudyConfig
-	Models    []*api.ModelInfo
-}
