+++ docs/CLI/katib-cli.md
@@ -1,0 +1,24 @@
+## katib-cli
+
+katib cli
+
+### Synopsis
+
+This is katib cli client using cobra framework
+
+### Options
+
+```
+  -h, --help            help for katib-cli
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli create](katib-cli_create.md)	 - Create a resource from a file
+* [katib-cli get](katib-cli_get.md)	 - Display one or many resources
+* [katib-cli pull](katib-cli_pull.md)	 - Pull a resource from a file or from stdin.
+* [katib-cli push](katib-cli_push.md)	 - Push a resource from a file or from stdin.
+* [katib-cli stop](katib-cli_stop.md)	 - Stop a resource
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
