+++ docs/CLI/katib-cli_create.md
@@ -1,0 +1,26 @@
+## katib-cli create
+
+Create a resource from a file
+
+### Synopsis
+
+YAML formats are accepted.
+
+### Options
+
+```
+  -h, --help   help for create
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli](katib-cli.md)	 - katib cli
+* [katib-cli create study](katib-cli_create_study.md)	 - Create a study from a file
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
