+++ docs/CLI/katib-cli_create_study.md
@@ -1,0 +1,30 @@
+## katib-cli create study
+
+Create a study from a file
+
+### Synopsis
+
+YAML formats are accepted.
+
+```
+katib-cli create study [flags]
+```
+
+### Options
+
+```
+  -f, --config string   File path of study config(required)
+  -h, --help            help for study
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli create](katib-cli_create.md)	 - Create a resource from a file
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
