+++ docs/CLI/katib-cli_get.md
@@ -1,0 +1,27 @@
+## katib-cli get
+
+Display one or many resources
+
+### Synopsis
+
+list of resorces comannd can display includes: studies, study, trials, trial, models, model
+
+### Options
+
+```
+  -h, --help   help for get
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli](katib-cli.md)	 - katib cli
+* [katib-cli get model](katib-cli_get_model.md)	 - Display Model Info
+* [katib-cli get studies](katib-cli_get_studies.md)	 - Display Study lnfo
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
