+++ docs/CLI/katib-cli_get_model.md
@@ -1,0 +1,30 @@
+## katib-cli get model
+
+Display Model Info
+
+### Synopsis
+
+Display Information of saved model
+
+```
+katib-cli get model [flags]
+```
+
+### Options
+
+```
+  -d, --detail   Display detail information of Model
+  -h, --help     help for model
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli get](katib-cli_get.md)	 - Display one or many resources
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
