+++ docs/CLI/katib-cli_get_studies.md
@@ -1,0 +1,29 @@
+## katib-cli get studies
+
+Display Study lnfo
+
+### Synopsis
+
+Display Information of a studies
+
+```
+katib-cli get studies [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for studies
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli get](katib-cli_get.md)	 - Display one or many resources
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
