+++ docs/CLI/katib-cli_pull.md
@@ -1,0 +1,26 @@
+## katib-cli pull
+
+Pull a resource from a file or from stdin.
+
+### Synopsis
+
+YAML or JSON formats are accepted.
+
+### Options
+
+```
+  -h, --help   help for pull
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli](katib-cli.md)	 - katib cli
+* [katib-cli pull studies](katib-cli_pull_studies.md)	 - Export a Study and its Models lnfo
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
