+++ docs/CLI/katib-cli_pull_studies.md
@@ -1,0 +1,30 @@
+## katib-cli pull studies
+
+Export a Study and its Models lnfo
+
+### Synopsis
+
+Export Information of a Study and its Models to yaml format
+
+```
+katib-cli pull studies [flags]
+```
+
+### Options
+
+```
+  -h, --help            help for studies
+  -o, --output string   File path to export
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli pull](katib-cli_pull.md)	 - Pull a resource from a file or from stdin.
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
