+++ docs/CLI/katib-cli_push.md
@@ -1,0 +1,26 @@
+## katib-cli push
+
+Push a resource from a file or from stdin.
+
+### Synopsis
+
+YAML or JSON formats are accepted.
+
+### Options
+
+```
+  -h, --help   help for push
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli](katib-cli.md)	 - katib cli
+* [katib-cli push study](katib-cli_push_study.md)	 - Push a Study Info and its Models from a file or from stdin
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
