+++ docs/CLI/katib-cli_push_study.md
@@ -1,0 +1,31 @@
+## katib-cli push study
+
+Push a Study Info and its Models from a file or from stdin
+
+### Synopsis
+
+Push a Study Info and its Models from a file or from stdin
+YAML formats are accepted.
+
+```
+katib-cli push study [flags]
+```
+
+### Options
+
+```
+  -f, --file string   File path of model config file
+  -h, --help          help for study
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli push](katib-cli_push.md)	 - Push a resource from a file or from stdin.
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
