+++ docs/CLI/katib-cli_stop.md
@@ -1,0 +1,26 @@
+## katib-cli stop
+
+Stop a resource
+
+### Synopsis
+
+Specify resource ID or Name.
+
+### Options
+
+```
+  -h, --help   help for stop
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli](katib-cli.md)	 - katib cli
+* [katib-cli stop study](katib-cli_stop_study.md)	 - Stop a study
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
