+++ docs/CLI/katib-cli_stop_study.md
@@ -1,0 +1,29 @@
+## katib-cli stop study
+
+Stop a study
+
+### Synopsis
+
+Stop study with study ID or study name
+
+```
+katib-cli stop study [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for study
+```
+
+### Options inherited from parent commands
+
+```
+  -s, --server string   katib manager API endpoint (default "localhost:6789")
+```
+
+### SEE ALSO
+
+* [katib-cli stop](katib-cli_stop.md)	 - Stop a resource
+
+###### Auto generated by spf13/cobra on 5-Jun-2018
